\name{fAsianOptions-package}


\alias{fAsianOptions-package}
\alias{fAsianOptions}


\docType{package}


\title{EBM and Asian Option Valuation}


\description{

    The Rmetrics "fAsianOptions" package is a collection of functions to 
    analyze and model Exponential Brownian Motion and to valuate Asian
    options. 
    
}


\details{

    \preformatted{
        Package:    \tab fAsianOptions\cr
        Type:       \tab Package\cr
        Version:    \tab R 3.0.1\cr
        Date:       \tab 2014\cr
        License:    \tab GPL Version 2 or later\cr
        Copyright:  \tab (c) 1999-2014 Rmetrics Association\cr
        Repository: \tab R-FORGE\cr
        URL:        \tab \url{https://www.rmetrics.org}
    }
    
}


\keyword{package}


